refactor: misc cleanup
