--- lib/constants.ts
@@ -1,3 +1,0 @@
-export const CONFIG_DEFAULT = 'presta.config.js'
-export const OUTPUT_STATIC_DIR = 'static'
-export const OUTPUT_DYNAMIC_PAGES_ENTRY = 'functions/presta.js'
