--- lib/devServerIcon.ts
@@ -1,19 +1,0 @@
-const style = [
-  'position: fixed',
-  'bottom: 24px',
-  'right: 24px',
-  'width: 32px',
-  'height: 32px',
-  'border-radius: 32px',
-  'background: white',
-  'color: #FF7A93',
-  'font-size: 20px',
-  'font-weight: bold',
-  'text-align: center',
-  'line-height: 31px',
-  'box-shadow: 0px 0px 1px rgba(0, 0, 0, 0.04), 0px 4px 8px rgba(0, 0, 0, 0.04), 0px 16px 24px rgba(0, 0, 0, 0.04), 0px 24px 32px rgba(0, 0, 0, 0.04)'
-]
-
-export const devServerIcon = `
-  <div style="${style.join(';')}">~</div>
-`
