+++ lib/load.ts
@@ -1,15 +1,77 @@
+import path from 'path'
+import fs from 'fs-extra'
 import c from 'ansi-colors'
 
-import { createCache } from './loadCache'
-
-const { NODE_ENV } = process.env
+import { log } from './log'
+import { getCurrentConfig, Env } from './config'
 
 const requests = {}
 const errors = {}
-export const loadCache = createCache('presta-load-cache')
+export const loadCache = createLoadCache('presta-load-cache')
 
-function log (str: string) {
-  if (NODE_ENV !== 'test') console.log(str)
+function writeFileCache (filepath: string, json: object) {
+  if (getCurrentConfig().env !== Env.PRODUCTION)
+    fs.writeFileSync(filepath, JSON.stringify(json), 'utf-8')
+}
+
+function readFileCache (filepath: string) {
+  if (getCurrentConfig().env === Env.PRODUCTION) return {}
+  if (!fs.existsSync(filepath)) fs.writeFileSync(filepath, '{}', 'utf-8')
+  return JSON.parse(fs.readFileSync(filepath))
+}
+
+export function createLoadCache (name: string, { dir = process.cwd() } = {}) {
+  const filename = '.' + name
+  const filepath = path.join(dir, filename)
+
+  let cache = readFileCache(filepath)
+
+  return {
+    get (key: string) {
+      const [value, expiration] = cache[key] || []
+
+      if (expiration !== null && Date.now() > expiration) {
+        delete cache[key]
+        writeFileCache(filepath, cache)
+        return undefined
+      } else {
+        return value
+      }
+    },
+    set (key: string, value: any, duration?: number) {
+      const expiration = duration ? Date.now() + duration : null
+      cache[key] = [value, expiration]
+
+      if (expiration) writeFileCache(filepath, cache)
+    },
+    clear (key: string) {
+      delete cache[key]
+      writeFileCache(filepath, cache)
+    },
+    clearAllMemory () {
+      for (const key of Object.keys(cache)) {
+        const [value, expiration] = cache[key] || []
+        if (!expiration) delete cache[key]
+      }
+    },
+    cleanup () {
+      cache = {}
+
+      // no persistent cache may have been created
+      try {
+        fs.unlinkSync(filepath)
+      } catch (e) {}
+    },
+    dump () {
+      const res = {}
+
+      for (const key of Object.keys(cache)) {
+        res[key] = cache[key][0]
+      }
+
+      return res
+    }
+  }
 }
 
 export function loadError (key: string, e: Error) {
